feat(rust): share system font registration with Node
